requirements.txt
Django>=4.2,<5.0
djangorestframework==3.14.0
drf-nested-routers==0.93.4
django-cors-headers==4.3.1
python-dotenv>=1.0.0
django-crispy-forms>=2.0
crispy-bootstrap5>=0.7 